+++ core/src/ch/jeda/ui/Button.java
@@ -97,7 +97,7 @@ public class Button extends Widget implements KeyDownListener, KeyUpListener, Po
     }
 
     @Override
-    public final boolean contains(final int x, final int y) {
+    public final boolean contains(final float x, final float y) {
         return style.contains(this, x, y);
     }
